+++ DefenderAPI/functions/Mde/alerts/New-MdAlert.ps1
@@ -8,14 +8,23 @@
 
     Scopes required (delegate auth): Alert.ReadWrite
 
+.PARAMETER ReportID
+    Report Id of the event
+
 .PARAMETER EventTime
     Time of the event as string, e.g. 2018-08-03T16:45:21.7115183Z
 
 .PARAMETER MachineID
     ID of the machine on which the event was identified
 
+.PARAMETER RecommendedAction
+    Recommended action for the Alert
+
 .PARAMETER Title
     Title of the Alert
+
+.PARAMETER Category
+    Category of the alert
 
 .PARAMETER Severity
     Severity of the alert.
@@ -23,17 +32,8 @@
 .PARAMETER Description
     Description of the Alert
 
-.PARAMETER RecommendedAction
-    Recommended action for the Alert
-
-.PARAMETER Category
-    Category of the alert
-
-.PARAMETER ReportID
-    Report Id of the event
-
 .EXAMPLE
-    PS C:\> New-MdAlert -Title $title -Severity $severity -Description $description -Category $category
+    PS C:\> New-MdAlert -Title $title -Category $category -Severity $severity -Description $description
 
     Create Alert based on specific Event
 
@@ -45,23 +45,15 @@
     param (
         [Parameter(ValueFromPipelineByPropertyName = $true, ParameterSetName = 'default')]
         [string]
+        $ReportID,
+
+        [Parameter(ValueFromPipelineByPropertyName = $true, ParameterSetName = 'default')]
+        [string]
         $EventTime,
 
         [Parameter(ValueFromPipelineByPropertyName = $true, ParameterSetName = 'default')]
         [string]
         $MachineID,
-
-        [Parameter(Mandatory = $true, ValueFromPipelineByPropertyName = $true, ParameterSetName = 'default')]
-        [string]
-        $Title,
-
-        [Parameter(Mandatory = $true, ValueFromPipelineByPropertyName = $true, ParameterSetName = 'default')]
-        [string]
-        $Severity,
-
-        [Parameter(Mandatory = $true, ValueFromPipelineByPropertyName = $true, ParameterSetName = 'default')]
-        [string]
-        $Description,
 
         [Parameter(ValueFromPipelineByPropertyName = $true, ParameterSetName = 'default')]
         [string]
@@ -69,37 +61,45 @@
 
         [Parameter(Mandatory = $true, ValueFromPipelineByPropertyName = $true, ParameterSetName = 'default')]
         [string]
+        $Title,
+
+        [Parameter(Mandatory = $true, ValueFromPipelineByPropertyName = $true, ParameterSetName = 'default')]
+        [string]
         $Category,
 
-        [Parameter(ValueFromPipelineByPropertyName = $true, ParameterSetName = 'default')]
+        [Parameter(Mandatory = $true, ValueFromPipelineByPropertyName = $true, ParameterSetName = 'default')]
         [string]
-        $ReportID
+        $Severity,
+
+        [Parameter(Mandatory = $true, ValueFromPipelineByPropertyName = $true, ParameterSetName = 'default')]
+        [string]
+        $Description
     )
     process {
 		$__mapping = @{
+            'ReportID' = 'Report ID'
             'EventTime' = 'Event Time'
             'MachineID' = 'Machine ID'
+            'RecommendedAction' = 'Recommended Action'
             'Title' = 'Title'
+            'Category' = 'Category'
             'Severity' = 'Severity'
             'Description' = 'Description'
-            'RecommendedAction' = 'Recommended Action'
-            'Category' = 'Category'
-            'ReportID' = 'Report ID'
 		}
 
 		$__param = @{
-			Body = $PSBoundParameters | ConvertTo-HashTable -Include @('EventTime','MachineID','Title','Severity','Description','RecommendedAction','Category','ReportID') -Mapping $__mapping
+			Body = $PSBoundParameters | ConvertTo-HashTable -Include @('ReportID','EventTime','MachineID','RecommendedAction','Title','Category','Severity','Description') -Mapping $__mapping
 			Query = $PSBoundParameters | ConvertTo-HashTable -Include @() -Mapping $__mapping
 			Header = $PSBoundParameters | ConvertTo-HashTable -Include @() -Mapping $__mapping
 			Path = 'alerts/createAlertByReference'
 			Method = 'post'
 			RequiredScopes = 'Alert.ReadWrite'
-			
+			Service = 'DefenderAPI.Endpoint'
 		}
 		
 		$__param += $PSBoundParameters | ConvertTo-HashTable -Include 'ErrorAction', 'WarningAction', 'Verbose'
 
-		try { Invoke-DefenderAPIRequest @__param }
+		try { Invoke-EntraRequest @__param }
 		catch { $PSCmdlet.ThrowTerminatingError($_) }
     }
 }