+++ DefenderAPI/functions/Mde/indicators/Get-MdIndicator.ps1
@@ -15,14 +15,14 @@
 .PARAMETER Select
     Selects which properties to include in the response, defaults to all.
 
-.PARAMETER Skip
-    Skips the first n results.
-
 .PARAMETER Filter
     Filters the results, using OData syntax.
 
 .PARAMETER Count
     Includes a count of the matching results in the response.
+
+.PARAMETER Skip
+    Skips the first n results.
 
 .EXAMPLE
     PS C:\> Get-MdIndicator
@@ -47,40 +47,40 @@
         $Select,
 
         [Parameter(ValueFromPipelineByPropertyName = $true, ParameterSetName = 'default')]
-        [int32]
-        $Skip,
-
-        [Parameter(ValueFromPipelineByPropertyName = $true, ParameterSetName = 'default')]
         [string]
         $Filter,
 
         [Parameter(ValueFromPipelineByPropertyName = $true, ParameterSetName = 'default')]
         [boolean]
-        $Count
+        $Count,
+
+        [Parameter(ValueFromPipelineByPropertyName = $true, ParameterSetName = 'default')]
+        [int32]
+        $Skip
     )
     process {
 		$__mapping = @{
             'Top' = '$top'
             'Orderby' = '$orderby'
             'Select' = '$select'
-            'Skip' = '$skip'
             'Filter' = '$filter'
             'Count' = '$count'
+            'Skip' = '$skip'
 		}
 
 		$__param = @{
 			Body = $PSBoundParameters | ConvertTo-HashTable -Include @() -Mapping $__mapping
-			Query = $PSBoundParameters | ConvertTo-HashTable -Include @('Top','Orderby','Select','Skip','Filter','Count') -Mapping $__mapping
+			Query = $PSBoundParameters | ConvertTo-HashTable -Include @('Top','Orderby','Select','Filter','Count','Skip') -Mapping $__mapping
 			Header = $PSBoundParameters | ConvertTo-HashTable -Include @() -Mapping $__mapping
 			Path = 'indicators'
 			Method = 'get'
 			
-			
+			Service = 'DefenderAPI.Endpoint'
 		}
 		
 		$__param += $PSBoundParameters | ConvertTo-HashTable -Include 'ErrorAction', 'WarningAction', 'Verbose'
 
-		try { Invoke-DefenderAPIRequest @__param }
+		try { Invoke-EntraRequest @__param }
 		catch { $PSCmdlet.ThrowTerminatingError($_) }
     }
 }