+++ DefenderAPI/functions/Mde/indicators/Remove-MdIndicator.ps1
@@ -36,12 +36,12 @@
 			Path = 'indicators/{IndicatorID}' -Replace '{IndicatorID}',$IndicatorID
 			Method = 'delete'
 			
-			
+			Service = 'DefenderAPI.Endpoint'
 		}
 		
 		$__param += $PSBoundParameters | ConvertTo-HashTable -Include 'ErrorAction', 'WarningAction', 'Verbose'
 
-		try { Invoke-DefenderAPIRequest @__param }
+		try { Invoke-EntraRequest @__param }
 		catch { $PSCmdlet.ThrowTerminatingError($_) }
     }
 }