+++ DefenderAPI/functions/Mde/machines/Get-MdMachineSoftware.ps1
@@ -38,12 +38,12 @@
 			Path = 'machines/{MachineID}/software' -Replace '{MachineID}',$MachineID
 			Method = 'get'
 			RequiredScopes = 'Software.Read'
-			
+			Service = 'DefenderAPI.Endpoint'
 		}
 		
 		$__param += $PSBoundParameters | ConvertTo-HashTable -Include 'ErrorAction', 'WarningAction', 'Verbose'
 
-		try { Invoke-DefenderAPIRequest @__param }
+		try { Invoke-EntraRequest @__param }
 		catch { $PSCmdlet.ThrowTerminatingError($_) }
     }
 }