+++ DefenderAPI/functions/Mde/machines/Set-MdMachineStartinvestigation.ps1
@@ -44,12 +44,12 @@
 			Path = 'machines/{MachineID}/startInvestigation' -Replace '{MachineID}',$MachineID
 			Method = 'post'
 			
-			
+			Service = 'DefenderAPI.Endpoint'
 		}
 		
 		$__param += $PSBoundParameters | ConvertTo-HashTable -Include 'ErrorAction', 'WarningAction', 'Verbose'
 
-		try { Invoke-DefenderAPIRequest @__param }
+		try { Invoke-EntraRequest @__param }
 		catch { $PSCmdlet.ThrowTerminatingError($_) }
     }
 }