+++ DefenderAPI/internal/classes/SubnetCategory.ps1
@@ -1,0 +1,9 @@
+﻿enum SubnetCategory {
+	Unknown = 0
+	Corporate = 1
+	Administrative = 2
+	Risky = 3
+	VPN = 4
+	CloudProvider = 5
+	Other = 6
+}