+++ DefenderAPI/internal/scripts/preimport.ps1
@@ -13,7 +13,6 @@ $moduleRoot = Split-Path (Split-Path $PSScriptRoot)
 # Load the strings used in messages
 "$moduleRoot\internal\scripts\strings.ps1"
 
-# Load classes
 (Get-ChildItem "$moduleRoot\internal\classes\*.ps1" -ErrorAction Ignore).FullName
 
 # Load Tab Expansion