+++ DefenderAPI/internal/tepp/service.tepp.ps1
@@ -1,3 +1,0 @@
-﻿Register-PSFTeppScriptblock -Name 'DefenderAPI.Service' -ScriptBlock {
-	(Get-DefenderAPIService).Name
-} -Global